+++ NEWS.md
@@ -1,5 +1,15 @@
 # News
 
+## Release V0.5.4
+
+- Added functionality to solve reliability problems with extremely small probabilities of failure using the Importance Sampling technique.
+- Isoprobabilistic transformation objects (`NatafTransformation` and `RosenblattTransformation`) are now broadcastable.
+
+```julia
+jointpdf.(NatafObject, XSamples)
+jointpdf.(RosenblattObject, XSamples)
+```
+
 ## Release V0.5.3
 
 - HLRF and iHLRF methods can now be started from any arbitrary point $x_{0}$.